Move from fmwork package to fmwork module
Simplify structure (and imports inside engine interfaces)
by removing the fmwork utilities package and instead implement a
single fmwork module (like previous versions of FMwork).
Engine interfaces (e.g., embed/tf, infer/vllm) can symlink to
fmwork.py directly and just import it.
